feat: page characters
