+++ src/main.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { ErrorBoundary } from './components/ErrorBoundary.tsx';
+import { ErrorBoundary } from './components/error/ErrorBoundary.tsx';
 import App from './App.tsx';
 import './styles/global.css';
 import { scribe } from './scribe.ts';
@@ -11,8 +11,8 @@ console.log(scribe);
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <ErrorBoundary>
-      <Backdop />
       <main className="main">
+        <Backdop />
         <App />
       </main>
     </ErrorBoundary>
