+++ src/pages/page.module.css
@@ -1,7 +1,5 @@
-.page__results {
-  width: 100%;
-  box-sizing: border-box;
-  padding: 0 1rem;
+.page__section {
+  margin: 1rem 0 1.5rem;
 }
 
 .page__btns {
